requirements.txt
aiosqlite==0.21.0
discord.py==2.5.2
matplotlib==3.9.2
msgspec==0.18.6